Fix no features build
