+++ src/autofloat/sub.rs
@@ -76,7 +76,7 @@ where
 #[cfg(test)]
 mod test {
 
-    use crate::autofloat::test::assert_autofloat_eq;
+    use crate::test::assert_autofloat_eq;
 
     use super::*;
