+++ src/lib.rs
@@ -101,3 +101,15 @@ where
     }
     result
 }
+
+#[cfg(test)]
+mod test {
+    /// Convenience macro for comparing `AutoFloats`s in full.
+    macro_rules! assert_autofloat_eq {
+        ($lhs:expr, $rhs:expr $(,)?) => {
+            assert_eq!($lhs.x, $rhs.x);
+            assert_eq!($lhs.dx, $rhs.dx);
+        };
+    }
+    pub(crate) use assert_autofloat_eq;
+}
